Fix endpoints and separated doPost servlet
